refactor: reorganize utils and add `retry` sub-module
