+++ chain-signatures/indexer-core/Cargo.toml
@@ -7,6 +7,7 @@ description = "Core traits and types for the MPC indexers"
 [dependencies]
 alloy-primitives.workspace = true
 alloy-sol-types.workspace = true
+backon.workspace = true
 # TODO: replace anyhow with thiserror
 anyhow.workspace = true
 async-trait.workspace = true
