+++ chain-signatures/node/src/rpc/mod.rs
@@ -10,7 +10,6 @@ use crate::metrics::requests::{record_request_latency_since, SignRequestStep};
 use crate::protocol::contract::primitives::{ParticipantMap, Participants};
 use crate::protocol::contract::RunningContractState;
 use crate::protocol::{Chain, IndexedSignRequest, ProtocolState};
-use crate::util::retry::{retry_rpc, RetryConfig};
 use enum_map::EnumMap;
 use std::collections::BTreeSet;
 use std::sync::Arc;
@@ -25,6 +24,7 @@ use cait_sith::FullSignature;
 use k256::{AffinePoint, Secp256k1};
 pub use mpc_contract::primitives::{Read, View};
 use mpc_primitives::{CheckpointDigest, Signature};
+use mpc_indexer_core::utils::retry::{RetryConfig, retry_rpc};
 pub use near::NearClient;
 
 use near_account_id::AccountId;
